+++ backend/api/posts.go
@@ -49,17 +49,22 @@ func GetPost(w http.ResponseWriter, r *http.Request) {
 	w.Write(jsonRes)
 }
 
-func DeletePost(w http.ResponseWriter, r *http.Request) {
+func DeletePost(w http.ResponseWriter, r *http.Request, user_id string) {
 	postId := path.Base(r.URL.Path)
 
-	db := database.Connect()
+	if AuthToPost(postId, user_id) {
+		db := database.Connect()
 
-	database.DeleteById(db, "posts", postId)
+		database.DeleteById(db, "posts", postId)
 
-	database.Disconnect(db)
+		database.Disconnect(db)
+	} else {
+		http.Error(w, "Unauthorized to remove this post", http.StatusUnauthorized)
+		return
+	}
 }
 
-func PatchPost(w http.ResponseWriter, r *http.Request) {
+func PatchPost(w http.ResponseWriter, r *http.Request, user_id string) {
 	postId := path.Base(r.URL.Path)
 	var post models.Posts
 	err := json.NewDecoder(r.Body).Decode(&post)
@@ -69,26 +74,41 @@ func PatchPost(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var columns []string
-	var values []string
-	if post.Title != "" {
-		columns = append(columns, "title")
-		values = append(values, post.Title)
+	if AuthToPost(postId, user_id) {
+		var columns []string
+		var values []string
+		if post.Title != "" {
+			columns = append(columns, "title")
+			values = append(values, post.Title)
+		}
+		if post.Description != "" {
+			columns = append(columns, "description")
+			values = append(values, post.Description)
+		}
+		table := "posts"
+
+		db := database.Connect()
+
+		database.PatchById(db, table, postId, columns, values)
+
+		database.Disconnect(db)
+	} else {
+		http.Error(w, "Unauthorized to update this post", http.StatusUnauthorized)
+		return
 	}
-	if post.Description != "" {
-		columns = append(columns, "description")
-		values = append(values, post.Description)
-	}
-	table := "posts"
-
-	db := database.Connect()
-
-	database.PatchById(db, table, postId, columns, values)
-
-	database.Disconnect(db)
 }
 
-func CreatePost(w http.ResponseWriter, r *http.Request) {
+func AuthToPost(postId string, userId string) bool {
+	db := database.Connect()
+	var post []models.Posts
+	database.GetRowById(db, &post, "posts", postId)
+
+	database.Disconnect(db)
+
+	return post[0].Poster == userId
+}
+
+func CreatePost(w http.ResponseWriter, r *http.Request, user_id string) {
 	var post models.Posts
 	err := json.NewDecoder(r.Body).Decode(&post)
